fix warnings

+++ project1-1/src/productionLine/CalcDispatcher.cpp
@@ -15,7 +15,7 @@ int CalcDispatcher::run() {
     }
     Log::append(Log::INFO, "CalcDispatcher", "stop");
     CalcRequest emptyCalcRequest;
-    for (int i = 0; i < this->clients.size(); i++) {
+    for (size_t i = 0; i < this->clients.size(); i++) {
         CalcDispatcherClient* client = this->clients[i];
         client->sendNetworkMessage(emptyCalcRequest);
     }
